% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dose_response_curve.R
\name{fit_ic50_curve}
\alias{fit_ic50_curve}
\title{Get Dose Response Curve}
\usage{
fit_ic50_curve(df, molarity)
}
\arguments{
\item{df}{The dataframe containing the dose-response data in the structure of testdata}

\item{molarity}{Any string to be assigned to the molarity value, for example "nm", "um", or "percent"}
}
\value{
The function returns the IC50 value, a dose-response plot as a ggplot object,
an interactive dose-response plot as a plotly object, and summary statistics
}
\description{
Dose response curve fitting and IC50 calculation using 4-parameter logistic model
}
